chore: extract transaction info components
